Use ingestion pipeline in Python code (#61)
---------
