+++ .changeset/short-ducks-drum.md
@@ -1,0 +1,5 @@
+---
+"create-llama": patch
+---
+
+Use ingestion pipeline for Python
